moved fredpox baseline to top-level dir; updated structure according to new template [WIP]
