+++ baselines/fedprox/fedprox/server.py
@@ -7,7 +7,7 @@ from hydra.utils import instantiate
 from omegaconf import DictConfig
 from torch.utils.data import DataLoader
 
-from flwr_baselines.publications.fedprox_mnist.models import test
+from fedprox.models import test
 
 def gen_evaluate_fn(
     testloader: DataLoader,
