+++ baselines/fedprox_/docs/FedProx_mnist.png
