--- baselines/flwr_baselines/flwr_baselines/publications/fedprox_mnist/models.py
@@ -1,197 +1,0 @@
-"""CNN model architecutre, training, and testing functions for MNIST."""
-
-
-from typing import List, Tuple
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch.nn.parameter import Parameter
-from torch.utils.data import DataLoader
-
-
-class Net(nn.Module):
-    """Convolutional Neural Network architecture.
-
-    As described in McMahan 2017 paper :
-
-    [Communication-Efficient Learning of Deep Networks from
-    Decentralized Data] (https://arxiv.org/pdf/1602.05629.pdf)
-    """
-
-    def __init__(self, num_classes: int) -> None:
-        super().__init__()
-        self.conv1 = nn.Conv2d(1, 32, 5, padding=1)
-        self.conv2 = nn.Conv2d(32, 64, 5, padding=1)
-        self.pool = nn.MaxPool2d(kernel_size=(2, 2), padding=1)
-        self.fc1 = nn.Linear(64 * 7 * 7, 512)
-        self.fc2 = nn.Linear(512, num_classes)
-
-    def forward(self, input_tensor: torch.Tensor) -> torch.Tensor:
-        """Forward pass of the CNN.
-
-        Parameters
-        ----------
-        x : torch.Tensor
-            Input Tensor that will pass through the network
-
-        Returns
-        -------
-        torch.Tensor
-            The resulting Tensor after it has passed through the network
-        """
-        output_tensor = F.relu(self.conv1(input_tensor))
-        output_tensor = self.pool(output_tensor)
-        output_tensor = F.relu(self.conv2(output_tensor))
-        output_tensor = self.pool(output_tensor)
-        output_tensor = torch.flatten(output_tensor, 1)
-        output_tensor = F.relu(self.fc1(output_tensor))
-        output_tensor = self.fc2(output_tensor)
-        return output_tensor
-
-
-class LogisticRegression(nn.Module):
-    """A network for logistic regression using a single fully connected layer.
-
-    As described in Lin 2020 paper :
-
-    [Federated Optimization in Heterogeneous Networks]
-    (https://arxiv.org/pdf/1812.06127.pdf)
-    """
-
-    def __init__(self, num_classes: int) -> None:
-        super().__init__()
-        self.fc = nn.Linear(28 * 28, num_classes)
-
-    def forward(self, input_tensor: torch.Tensor) -> torch.Tensor:
-        """Forward pass.
-
-        Parameters
-        ----------
-        x : torch.Tensor
-            Input Tensor that will pass through the network
-
-        Returns
-        -------
-        torch.Tensor
-            The resulting Tensor after it has passed through the network
-        """
-        output_tensor = self.fc(torch.flatten(input_tensor, 1))
-        return output_tensor
-
-
-def train(  # pylint: disable=too-many-arguments
-    net: nn.Module,
-    trainloader: DataLoader,
-    device: torch.device,
-    epochs: int,
-    learning_rate: float,
-    proximal_mu: float,
-) -> None:
-    """Train the network on the training set.
-
-    Parameters
-    ----------
-    net : nn.Module
-        The neural network to train.
-    trainloader : DataLoader
-        The DataLoader containing the data to train the network on.
-    device : torch.device
-        The device on which the model should be trained, either 'cpu' or 'cuda'.
-    epochs : int
-        The number of epochs the model should be trained for.
-    learning_rate : float
-        The learning rate for the SGD optimizer.
-    proximal_mu : float
-        Parameter for the weight of the proximal term.
-    """
-    criterion = torch.nn.CrossEntropyLoss()
-    optimizer = torch.optim.SGD(net.parameters(), lr=learning_rate, weight_decay=0.001)
-    global_params = [val.detach().clone() for val in net.parameters()]
-    net.train()
-    for _ in range(epochs):
-        net = _train_one_epoch(
-            net, global_params, trainloader, device, criterion, optimizer, proximal_mu
-        )
-
-
-def _train_one_epoch(  # pylint: disable=too-many-arguments
-    net: nn.Module,
-    global_params: List[Parameter],
-    trainloader: DataLoader,
-    device: torch.device,
-    criterion: torch.nn.CrossEntropyLoss,
-    optimizer: torch.optim.Adam,
-    proximal_mu: float,
-) -> nn.Module:
-    """Train for one epoch.
-
-    Parameters
-    ----------
-    net : nn.Module
-        The neural network to train.
-    global_params : List[Parameter]
-        The parameters of the global model (from the server).
-    trainloader : DataLoader
-        The DataLoader containing the data to train the network on.
-    device : torch.device
-        The device on which the model should be trained, either 'cpu' or 'cuda'.
-    criterion : torch.nn.CrossEntropyLoss
-        The loss function to use for training
-    optimizer : torch.optim.Adam
-        The optimizer to use for training
-    proximal_mu : float
-        Parameter for the weight of the proximal term.
-
-    Returns
-    -------
-    nn.Module
-        The model that has been trained for one epoch.
-    """
-    for images, labels in trainloader:
-        images, labels = images.to(device), labels.to(device)
-        optimizer.zero_grad()
-        proximal_term = 0.0
-        for local_weights, global_weights in zip(net.parameters(), global_params):
-            proximal_term += (local_weights - global_weights).norm(2)
-        loss = criterion(net(images), labels) + (proximal_mu / 2) * proximal_term
-        loss.backward()
-        optimizer.step()
-    return net
-
-
-def test(
-    net: nn.Module, testloader: DataLoader, device: torch.device
-) -> Tuple[float, float]:
-    """Evaluate the network on the entire test set.
-
-    Parameters
-    ----------
-    net : nn.Module
-        The neural network to test.
-    testloader : DataLoader
-        The DataLoader containing the data to test the network on.
-    device : torch.device
-        The device on which the model should be tested, either 'cpu' or 'cuda'.
-
-    Returns
-    -------
-    Tuple[float, float]
-        The loss and the accuracy of the input model on the given data.
-    """
-    criterion = torch.nn.CrossEntropyLoss()
-    correct, total, loss = 0, 0, 0.0
-    net.eval()
-    with torch.no_grad():
-        for images, labels in testloader:
-            images, labels = images.to(device), labels.to(device)
-            outputs = net(images)
-            loss += criterion(outputs, labels).item()
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-    if len(testloader.dataset) == 0:
-        raise ValueError("Testloader can't be 0, exiting...")
-    loss /= len(testloader.dataset)
-    accuracy = correct / total
-    return loss, accuracy
